pyproject.toml
[project]
name = "reflections"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi>=0.115.0,<1.0.0",
    "uvicorn[standard]>=0.30.0,<1.0.0",
    "pydantic-ai>=0.0.20",
    "httpx>=0.27.0,<1.0.0",
    "sqlalchemy>=2.0.0,<3.0.0",
    "alembic>=1.13.0,<2.0.0",
    "psycopg[binary]>=3.2.0,<4.0.0",
    "sentence-transformers>=3.0.0,<4.0.0",
    "pytest>=8.0.0,<9.0.0",
    "requests>=2.32.0,<3.0.0",
    "pydantic-settings>=2.6.0,<3.0.0",
]

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.1"
ruff = "^0.8.4"

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 88
target-version = "py312"
extend-exclude = ["apps/web"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.ruff.format]
quote-style = "double"

[tool.poetry]
packages = [{include = "reflections", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
